Task list of SEO Improvements5 Tasks
Add a H1 heading to this page.
Very important!

Use good headings on the page.
Very important!

This page has only very few internal links. If it's a welcome page, consider removing it.
Very important!

The page title should be better suited to the content of the page.
Important!

Improve the page response time.
Tip!